[Feature #1549] Move TermDto into a new package to prevent confusion with a planned new DTO representing term detail.
The existing TermDto represents a term in a list of terms and contains only a subset of its attributes.
