+++ src/main/java/cz/cvut/kbss/termit/service/business/readonly/ReadOnlyTermService.java
@@ -1,6 +1,6 @@
 package cz.cvut.kbss.termit.service.business.readonly;
 
-import cz.cvut.kbss.termit.dto.TermDto;
+import cz.cvut.kbss.termit.dto.listing.TermDto;
 import cz.cvut.kbss.termit.dto.readonly.ReadOnlyTerm;
 import cz.cvut.kbss.termit.model.Term;
 import cz.cvut.kbss.termit.model.Vocabulary;
