+++ src/test/java/cz/cvut/kbss/termit/service/business/readonly/ReadOnlyTermServiceTest.java
@@ -1,6 +1,6 @@
 package cz.cvut.kbss.termit.service.business.readonly;
 
-import cz.cvut.kbss.termit.dto.TermDto;
+import cz.cvut.kbss.termit.dto.listing.TermDto;
 import cz.cvut.kbss.termit.dto.TermInfo;
 import cz.cvut.kbss.termit.dto.readonly.ReadOnlyTerm;
 import cz.cvut.kbss.termit.environment.Generator;
